move getdata/putdata tests to same file
